Move graph code into helper
